Move prints to methods
